automarking: Reorganize testers part 1
