--- lib/automated_tests/examples/python/GreetPeople.py
@@ -1,4 +1,0 @@
-#!/usr/bin/env python
-
-def greet(name):
-    return "Hello " + name;
