--- lib/automated_tests/examples/python/args.in
@@ -1,4 +1,0 @@
-sample_files/python_sample/py_suite_1.py false
-sample_files/python_sample/py_suite_2.py false
-sample_files/python_sample/py_suite_3.py true
-sample_files/python_sample/py_suite_1.py false
