--- lib/automated_tests/examples/python/hello.py
@@ -1,7 +1,0 @@
-#!/usr/bin/env python
-
-def greet(name):
-    return "Hello " + name;
-    
-def goodbye(name):
-    return "Goodbye " + name
