--- lib/automated_tests/examples/racket/args.in
@@ -1,1 +1,0 @@
-sample_files/racket_sample/rkt_suite_1.rkt	false
