--- lib/automated_tests/examples/racket/code.rkt
@@ -1,11 +1,0 @@
-#lang racket
-
-(provide gt)
-(provide lt)
-
-
-(define (gt a b)
-	(> a b))
-	 
-(define (lt a b)
-	(< a b))
