--- lib/automated_tests/examples/ruby/Hello.rb
@@ -1,13 +1,0 @@
-def greet(x)
-  return "Hello #{x}"
-end
-
-def goodbye(x)
-  return "Goodbye x"
-end
-
-def greetly(x)
-  return 1/0
-end
-
-#puts greet("Hello World!")
