--- lib/automated_tests/examples/ruby/args.in
@@ -1,4 +1,0 @@
-sample_files/ruby_sample/rb_suite_1.rb false
-sample_files/ruby_sample/rb_suite_2.rb false
-sample_files/ruby_sample/rb_suite_3.rb true
-sample_files/ruby_sample/rb_suite_1.rb false
