--- lib/automated_tests/examples/shell/args.in
@@ -1,1 +1,0 @@
-shell_suite_1 false
