--- lib/automated_tests/examples/shell/expected_1.out
@@ -1,3 +1,0 @@
-Hello Bob
-Hello Mark
-Hello Frank
