--- lib/automated_tests/examples/shell/expected_2.out
@@ -1,3 +1,0 @@
-Goodbye Bob
-Goodbye Mark
-Goodbye George
