--- lib/automated_tests/examples/shell/input_1.in
@@ -1,3 +1,0 @@
-Bob
-Mark
-Frank
